+++ bin/cli.js
@@ -1,4 +1,3 @@
-'use strict';
 const path = require('path');
 
 const cli = require(path.resolve(__dirname, '../server/components/cli/index'));