+++ server/components/cli/index.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const chalk = require('chalk');
 const figlet = require('figlet');
 const inquirer = require('inquirer');