+++ server/components/cli/manga.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const fs = require('fs');
 const chalk = require('chalk');
 const folder = require('./folder');