+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGE LOG
 
+## v0.2.0
+
+* Implement server-side rendering, react-hot-reload
+
 ## v0.1.0
 
 * Data models: Author, Manga, Page
-* CLI tool for importing data to mysql
-* Client feature: Manga Dashboard (listing authors, mangas, image viewer)
+* CLI tool for importing data to MySQL
+* Client feature: Manga Dashboard (listing authors, mangas and manga viewer)